refactor: factor out shared courseMetadataBase.factory
