+++ src/shared/data/__factories__/index.js
@@ -1,3 +1,4 @@
 import './block.factory';
 import './courseBlocks.factory';
+import './courseMetadataBase.factory';
 import './tab.factory';
