+++ src/shared/data/__factories__/tab.factory.js
@@ -3,14 +3,16 @@ import { Factory } from 'rosie'; // eslint-disable-line import/no-extraneous-dep
 Factory.define('tab')
   .option('courseId', 'course-v1:edX+DemoX+Demo_Course')
   .option('path', 'course/')
+  .option('host', 'http://localhost:18000')
   .attrs({
     title: 'Course',
     priority: 0,
     slug: 'courseware',
     type: 'courseware',
   })
+  .attr('tab_id', ['slug'], (slug) => slug)
   .attr(
     'url',
-    ['courseId', 'path'],
-    (courseId, path) => `/courses/${courseId}/${path}`,
+    ['courseId', 'path', 'host'],
+    (courseId, path, host) => `${host}/courses/${courseId}/${path}`,
   );
